Unify and enrich tests
